hardcoded styling for dob picker
