+++ src/app/_layout.tsx
@@ -64,7 +64,7 @@ function Providers({ children }: { children: React.ReactNode }) {
   return (
     <GestureHandlerRootView
       style={styles.container}
-      className={theme.dark ? `dark` : undefined}
+      // className={theme.dark ? `dark` : undefined} //disable dark mode
     >
       <KeyboardProvider>
         <ThemeProvider value={theme}>
