+++ src/app/index.tsx
@@ -1,7 +1,7 @@
 import { Redirect } from 'expo-router';
 
 const Index = () => {
-  return <Redirect href={{ pathname: '/test' }} />;
+  return <Redirect href={{ pathname: '/_sitemap' }} />;
 };
 
 export default Index;
